refactor: DFS 탐색 관련 모델들을 별도의 패키지로 분리
