+++ src/test/kotlin/omok/model/search/HorizontalDfsTest.kt
@@ -1,5 +1,6 @@
-package omok.model
+package omok.model.search
 
+import omok.model.Color
 import omok.model.fixture.createPlayingBoard
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
